+++ tools/septentrio/septentrio_gnss/serial_comm.py
